Refactor linear modules
